+++ jms-db-domain/src/main/java/com/jms/domain/db/SectorMember.java
@@ -1,5 +1,5 @@
 package com.jms.domain.db;
-// Generated 2015-6-16 15:01:20 by Hibernate Tools 3.2.2.GA
+// Generated 2015-6-25 9:56:24 by Hibernate Tools 3.2.2.GA
 
 
 import javax.persistence.AttributeOverride;
@@ -26,7 +26,7 @@ public class SectorMember  implements java.io.Serializable {
      private Roles roles;
      private Users users;
      private Sector sector;
-     private Integer isprimary;
+     private Long isprimary;
 
     public SectorMember() {
     }
@@ -37,7 +37,7 @@ public class SectorMember  implements java.io.Serializable {
         this.users = users;
         this.sector = sector;
     }
-    public SectorMember(SectorMemberId id, Roles roles, Users users, Sector sector, Integer isprimary) {
+    public SectorMember(SectorMemberId id, Roles roles, Users users, Sector sector, Long isprimary) {
        this.id = id;
        this.roles = roles;
        this.users = users;
@@ -86,11 +86,11 @@ public class SectorMember  implements java.io.Serializable {
     }
     
     @Column(name="ISPRIMARY")
-    public Integer getIsprimary() {
+    public Long getIsprimary() {
         return this.isprimary;
     }
     
-    public void setIsprimary(Integer isprimary) {
+    public void setIsprimary(Long isprimary) {
         this.isprimary = isprimary;
     }
