+++ jms-repositories/src/main/java/com/jms/repositories/company/CompanyRepository.java
@@ -2,13 +2,13 @@ package com.jms.repositories.company;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
 import com.jms.domain.db.Company;
 
-
 @Repository
-public interface CompanyRepository extends CrudRepository<Company, Integer> {
+public interface CompanyRepository extends CrudRepository<Company, Long> {
 	
 	public Company findByCompanyName(String companyName);
+	
+
 	
 }
