+++ jms-ws-domain/src/main/java/com/jms/domain/ws/WSProvince.java
@@ -2,12 +2,12 @@ package com.jms.domain.ws;
 
 public class WSProvince implements java.io.Serializable{
 	
-	private int idProvince;
+	private Long idProvince;
 	private String province;
-	public int getIdProvince() {
+	public Long getIdProvince() {
 		return idProvince;
 	}
-	public void setIdProvince(int idProvince) {
+	public void setIdProvince(Long idProvince) {
 		this.idProvince = idProvince;
 	}
 	public String getProvince() {
